+++ apps/web/lib/constants/animations.ts
@@ -1,7 +1,7 @@
 export const fadeInUp = {
   initial: { opacity: 0, y: 20 },
   animate: { opacity: 1, y: 0 },
-  transition: { duration: 0.5 },
+  transition: { duration: 0.52 },
 };
 
 export const staggerChildren = {
@@ -19,7 +19,7 @@ export const fadeInUpAnimation = {
   initial: { opacity: 0, y: 20 },
   whileInView: { opacity: 1, y: 0 },
   viewport: { once: true },
-  transition: { duration: 0.6 }
+  transition: { duration: 0.52 }
 };
 
 export const fadeInUpVariants = {
