+++ apps/web/lib/mock-data/mock-projects.ts
@@ -1,21 +1,4 @@
-export interface Tag {
-  id: string;
-  text: string;
-}
-
-export interface Project {
-  id: string;
-  image: string;
-  category: string;
-  title: string;
-  description: string;
-  currentAmount: number;
-  targetAmount: number;
-  investors: number;
-  minInvestment: number;
-  percentageComplete: number;
-  tags: Tag[];
-}
+import { Project } from "~/lib/types/home.types";
 
 export const projects: Project[] = [
   {
